requirements.txt
gunicorn
python-dotenv

flask

pydantic
pydantic[email]

